+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - core.DataFrameTableModel: stop using QVariant
+- core.DataFrameTableModel.__init__: set dataFrame after connecting signals
 - core.DataFrameTableModel.headerData() defaults to Horizontal orientation
 
 ## [0.1.2] - 2024-10-01